Move metric sending into catabalancer package
