+++ events/events_test.go
@@ -1,10 +1,8 @@
 package events
 
 import (
-	"encoding/json"
 	"testing"
 
-	"github.com/livepeer/catalyst-api/balancer/catabalancer"
 	"github.com/stretchr/testify/require"
 )
 
@@ -47,43 +45,4 @@ func TestItFailsBadShapes(t *testing.T) {
 		_, err := Unmarshal(payload)
 		require.Error(t, err)
 	}
-}
-
-func TestItCanMarshalAndUnMarshalStreamIDs(t *testing.T) {
-	n := catabalancer.NodeUpdateEvent{}
-	n.SetStreams([]string{"noningest1", "noningest2"}, []string{"ingest1", "ingest2"})
-	jsonBytes, err := json.Marshal(n)
-	require.NoError(t, err)
-
-	var n2 catabalancer.NodeUpdateEvent
-	require.NoError(t, json.Unmarshal(jsonBytes, &n2))
-
-	require.Equal(t, []string{"noningest1", "noningest2"}, n2.GetStreams())
-	require.Equal(t, []string{"ingest1", "ingest2"}, n2.GetIngestStreams())
-}
-
-func TestItCanMarshalAndUnMarshalStreamIDsWithNoIngestStreams(t *testing.T) {
-	n := catabalancer.NodeUpdateEvent{}
-	n.SetStreams([]string{"noningest1", "noningest2"}, []string{})
-	jsonBytes, err := json.Marshal(n)
-	require.NoError(t, err)
-
-	var n2 catabalancer.NodeUpdateEvent
-	require.NoError(t, json.Unmarshal(jsonBytes, &n2))
-
-	require.Equal(t, []string{"noningest1", "noningest2"}, n2.GetStreams())
-	require.Equal(t, []string{}, n2.GetIngestStreams())
-}
-
-func TestItCanMarshalAndUnMarshalStreamIDsWithNoNonIngestStreams(t *testing.T) {
-	n := catabalancer.NodeUpdateEvent{}
-	n.SetStreams([]string{}, []string{"ingest1", "ingest2"})
-	jsonBytes, err := json.Marshal(n)
-	require.NoError(t, err)
-
-	var n2 catabalancer.NodeUpdateEvent
-	require.NoError(t, json.Unmarshal(jsonBytes, &n2))
-
-	require.Equal(t, []string{}, n2.GetStreams())
-	require.Equal(t, []string{"ingest1", "ingest2"}, n2.GetIngestStreams())
 }
